lock screen at specific position

+++ main.h
@@ -220,7 +220,7 @@ void display_progress_bar(void);
 u16 decrement_progress_bar(void);
 void bite_soldier(u16 x, u16 y);
 Sprite *get_sprite(u16 offset);
-void lock_screen(byte state);
+void lock_screen(u16 state);
 void fill_bottom_row(void);
 void finish_level(u16 i);
 void fade_in(u16 fade);
